move crypto fn to utils and split it into modules
